Starting to refactor and get more organized
