+++ client/src/components/overview/main-section/mainsection.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import styled from 'styled-components';
-import ProductImage from './productimage.jsx';
+import ProductImage from './product-image/productimage.jsx';
 import ProductInfo from './product-info/productinfo.jsx';
 
 function MainSection() {
